sensors.c: return curr from lowpass() if prev=NAN
Problem: Correct behavior of the lowpass filter depends on having a
valid previous value. However, the first time we want to read from a
sensor we cannot have a valid previous value. In this case, our best bet
is just to use the raw value. However, the current lowpass function
cannot return the raw value unless you pass in filterConstant=0. This
requires keeping track of first-time calls anyway, which is what we
wanted to avoid in the first place.

Instead, make lowpass() return the raw value if prevValue is set to NaN.
Then, the first time call is seamless as long as prevValue is set to
NaN.