Add scheduling VI functions
These functions are key for avoiding race-condition-like
bugs with our planned FreeRTOS scheduling setup. All
individual control loop iterations should be wrapped in
transactions, and before entering a transaction the loop
should check if the mode has been restarted. If it has, the
loop is no longer valid and the mode should reset to the
first-time mode setup.